ProfessorTransformer implemented

+++ spd-domain/src/main/java/br/com/spd/domain/Professor.java
@@ -1,7 +1,12 @@
 package br.com.spd.domain;
 
 import java.io.Serializable;
-import javax.persistence.*;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
 
 
 /**
@@ -14,7 +19,7 @@ public class Professor implements Serializable, Domain {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	private Integer id;
+	private Long id;
 
 	//bi-directional many-to-one association to Pessoa
 	@ManyToOne
@@ -29,11 +34,11 @@ public class Professor implements Serializable, Domain {
 	public Professor() {
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
